COMMERCE-7694 Rename
